Proof-of-concept implementation of semistructured merge
This is more a sketch of how it could be done quick&dirty than an actual
implementation.

Like in the fstmerge implementation, temporary files are created with
the content of the methods that should be merged. Then we just use our
existing LinebasedStrategy to write the merge result to another
temporary file. This content is just passed to the extendJ astnode.

Extendj had to be modified so that its pretty-printer uses the
precomputed String (if it's present) rather than traversing the subtree.

To integrate this branch with our main branches, we would have to at
least add an additional merge mode. Maybe also an additional strategy.

There are also several things that could be improved:
- as for the input, instead of just prettyprinting the methods into
temporary files, we could extract the code from the original files
using lineNumbers (I think we can get them from extendJ)
- as for the output, we could think about passing file pointers rather
than strings into extendJ in order to print the merged result

Signed-off-by: Olaf Lessenich <lessenic@fim.uni-passau.de>