Create mode semistructured.
The new mode can be used by supplying '-m semistructured' via command
line. This still needs extensive testing (there is no testsuite yet).

A switch for this feature had to be introduced both into the
MergeContext, as well as the ASTNodeArtifact. I'm not exactly happy with
this solution because of this redundancy, but it at least provides a
working integration of both structured and semistructured approaches.
However, this is a candidate for minor refactoring in the future.

Signed-off-by: Olaf Lessenich <lessenic@fim.uni-passau.de>