Fix wrong behavior in semistructured merge
Reporting all content nodes as matching with each other is a bug:
We miss insertion-deletion conflicts here as we do not distinguish the
version that comtains the insertion from the base version.

However, if we report whether they really match (i.e., are equal), they
are treated as conflicting during merging.

This patch provides a workaround for the situation: During merging, we
treat (not matching) content nodes different from (not matching) regular
nodes. While regular nodes end up in a conflict, content nodes still can
be merged. Therefore, we apply a MergeOperation in these cases.

Signed-off-by: Olaf Lessenich <lessenic@fim.uni-passau.de>